backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
asyncpg
python-dotenv 
pydantic
beautifulsoup4 
lxml
pdfplumber 
PyMuPDF 
openai 
requests
google-genai
pgvector